requirements.txt
pillow==10.3.0
matplotlib==3.9.0
tensorflow==2.15.1
scikit-learn==1.5.0
keras==2.15.0